+++ opentelemetry-sdk/src/logs/log_processor.rs
@@ -1,14 +1,10 @@
 use crate::{
     export::logs::{ExportResult, LogBatch, LogExporter},
     logs::{LogError, LogRecord, LogResult},
-    runtime::{RuntimeChannel, TrySend},
     Resource,
 };
-use futures_channel::oneshot;
-use futures_util::{
-    future::{self, Either},
-    {pin_mut, stream, StreamExt as _},
-};
+use std::sync::mpsc::{self, RecvTimeoutError, SyncSender};
+
 #[cfg(feature = "spec_unstable_logs_enabled")]
 use opentelemetry::logs::Severity;
 use opentelemetry::{otel_debug, otel_error, otel_warn, InstrumentationScope};
@@ -19,7 +15,19 @@ use std::{
     fmt::{self, Debug, Formatter},
     str::FromStr,
     sync::Arc,
+    thread,
     time::Duration,
+    time::Instant,
+};
+
+#[cfg(feature = "experimental_logs_batch_log_processor_with_async_runtime")]
+use crate::runtime::{RuntimeChannel, TrySend};
+#[cfg(feature = "experimental_logs_batch_log_processor_with_async_runtime")]
+use futures_channel::oneshot;
+#[cfg(feature = "experimental_logs_batch_log_processor_with_async_runtime")]
+use futures_util::{
+    future::{self, Either},
+    {pin_mut, stream, StreamExt as _},
 };
 
 /// Delay interval between two consecutive exports.
@@ -151,13 +159,40 @@ impl<T: LogExporter> LogProcessor for SimpleLogProcessor<T> {
     }
 }
 
-/// A [`LogProcessor`] that asynchronously buffers log records and reports
-/// them at a pre-configured interval.
-pub struct BatchLogProcessor<R: RuntimeChannel> {
-    message_sender: R::Sender<BatchMessage>,
+/// Messages sent between application thread and batch log processor's work thread.
+#[allow(clippy::large_enum_variant)]
+#[derive(Debug)]
+enum BatchMessage {
+    /// Export logs, usually called when the log is emitted.
+    ExportLog(Box<(LogRecord, InstrumentationScope)>),
+    /// Flush the current buffer to the backend, it can be triggered by
+    /// pre configured interval or a call to `force_push` function.
+    // Flush(Option<oneshot::Sender<ExportResult>>),
+    /// ForceFlush flushes the current buffer to the backend.
+    ForceFlush(mpsc::SyncSender<ExportResult>),
+    /// Shut down the worker thread, push all logs in buffer to the backend.
+    Shutdown(mpsc::SyncSender<ExportResult>),
+    /// Set the resource for the exporter.
+    SetResource(Arc<Resource>),
 }
 
-impl<R: RuntimeChannel> Debug for BatchLogProcessor<R> {
+/// A [`LogProcessor`] that buffers log records and reports
+/// them at a pre-configured interval from a dedicated background thread.
+pub struct BatchLogProcessor {
+    message_sender: SyncSender<BatchMessage>,
+    handle: Mutex<Option<thread::JoinHandle<()>>>,
+    forceflush_timeout: Duration,
+    shutdown_timeout: Duration,
+    is_shutdown: AtomicBool,
+
+    // Track dropped logs - we'll log this at shutdown
+    dropped_logs_count: AtomicUsize,
+
+    // Track the maximum queue size that was configured for this processor
+    max_queue_size: usize,
+}
+
+impl Debug for BatchLogProcessor {
     fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
         f.debug_struct("BatchLogProcessor")
             .field("message_sender", &self.message_sender)
@@ -165,12 +200,333 @@ impl<R: RuntimeChannel> Debug for BatchLogProcessor<R> {
     }
 }
 
-impl<R: RuntimeChannel> LogProcessor for BatchLogProcessor<R> {
+impl LogProcessor for BatchLogProcessor {
     fn emit(&self, record: &mut LogRecord, instrumentation: &InstrumentationScope) {
-        let result = self.message_sender.try_send(BatchMessage::ExportLog((
-            record.clone(),
-            instrumentation.clone(),
-        )));
+        // noop after shutdown
+        if self.is_shutdown.load(std::sync::atomic::Ordering::Relaxed) {
+            otel_warn!(
+                name: "BatchLogProcessor.Emit.ProcessorShutdown",
+                message = "BatchLogProcessor has been shutdown. No further logs will be emitted."
+            );
+            return;
+        }
+
+        let result = self
+            .message_sender
+            .try_send(BatchMessage::ExportLog(Box::new((
+                record.clone(),
+                instrumentation.clone(),
+            ))));
+
+        // TODO - Implement throttling to prevent error flooding when the queue is full or closed.
+        if result.is_err() {
+            // Increment dropped logs count. The first time we have to drop a log,
+            // emit a warning.
+            if self.dropped_logs_count.fetch_add(1, Ordering::Relaxed) == 0 {
+                otel_warn!(name: "BatchLogProcessor.LogDroppingStarted",
+                    message = "BatchLogProcessor dropped a LogRecord due to queue full/internal errors. No further log will be emitted for further drops until Shutdown. During Shutdown time, a log will be emitted with exact count of total logs dropped.");
+            }
+        }
+    }
+
+    fn force_flush(&self) -> LogResult<()> {
+        if self.is_shutdown.load(std::sync::atomic::Ordering::Relaxed) {
+            return LogResult::Err(LogError::Other(
+                "BatchLogProcessor is already shutdown".into(),
+            ));
+        }
+        let (sender, receiver) = mpsc::sync_channel(1);
+        self.message_sender
+            .try_send(BatchMessage::ForceFlush(sender))
+            .map_err(|err| LogError::Other(err.into()))?;
+
+        receiver
+            .recv_timeout(self.forceflush_timeout)
+            .map_err(|err| {
+                if err == RecvTimeoutError::Timeout {
+                    LogError::ExportTimedOut(self.forceflush_timeout)
+                } else {
+                    LogError::Other(err.into())
+                }
+            })?
+    }
+
+    fn shutdown(&self) -> LogResult<()> {
+        // test and set is_shutdown flag if it is not set
+        if self
+            .is_shutdown
+            .swap(true, std::sync::atomic::Ordering::Relaxed)
+        {
+            otel_warn!(
+                name: "BatchLogProcessor.Shutdown.ProcessorShutdown",
+                message = "BatchLogProcessor has been shutdown. No further logs will be emitted."
+            );
+            return LogResult::Err(LogError::AlreadyShutdown(
+                "BatchLogProcessor is already shutdown".into(),
+            ));
+        }
+
+        let dropped_logs = self.dropped_logs_count.load(Ordering::Relaxed);
+        let max_queue_size = self.max_queue_size;
+        if dropped_logs > 0 {
+            otel_warn!(
+                name: "BatchLogProcessor.LogsDropped",
+                dropped_logs_count = dropped_logs,
+                max_queue_size = max_queue_size,
+                message = "Logs were dropped due to a queue being full or other error. The count represents the total count of log records dropped in the lifetime of this BatchLogProcessor. Consider increasing the queue size and/or decrease delay between intervals."
+            );
+        }
+
+        let (sender, receiver) = mpsc::sync_channel(1);
+        self.message_sender
+            .try_send(BatchMessage::Shutdown(sender))
+            .map_err(|err| LogError::Other(err.into()))?;
+
+        receiver
+            .recv_timeout(self.shutdown_timeout)
+            .map(|_| {
+                // join the background thread after receiving back the shutdown signal
+                if let Some(handle) = self.handle.lock().unwrap().take() {
+                    handle.join().unwrap();
+                }
+                LogResult::Ok(())
+            })
+            .map_err(|err| match err {
+                RecvTimeoutError::Timeout => {
+                    otel_error!(
+                        name: "BatchLogProcessor.Shutdown.Timeout",
+                        message = "BatchLogProcessor shutdown timing out."
+                    );
+                    LogError::ExportTimedOut(self.shutdown_timeout)
+                }
+                _ => {
+                    otel_error!(
+                        name: "BatchLogProcessor.Shutdown.Error",
+                        error = format!("{}", err)
+                    );
+                    LogError::Other(err.into())
+                }
+            })?
+    }
+
+    fn set_resource(&self, resource: &Resource) {
+        let resource = Arc::new(resource.clone());
+        let _ = self
+            .message_sender
+            .try_send(BatchMessage::SetResource(resource));
+    }
+}
+
+impl BatchLogProcessor {
+    pub(crate) fn new(mut exporter: Box<dyn LogExporter>, config: BatchConfig) -> Self {
+        let (message_sender, message_receiver) = mpsc::sync_channel(config.max_queue_size);
+        let max_queue_size = config.max_queue_size;
+
+        let handle = thread::Builder::new()
+            .name("OpenTelemetry.Logs.BatchProcessor".to_string())
+            .spawn(move || {
+                let mut last_export_time = Instant::now();
+                let mut logs = Vec::new();
+                logs.reserve(config.max_export_batch_size);
+
+                loop {
+                    let remaining_time_option = config
+                        .scheduled_delay
+                        .checked_sub(last_export_time.elapsed());
+                    let remaining_time = match remaining_time_option {
+                        Some(remaining_time) => remaining_time,
+                        None => config.scheduled_delay,
+                    };
+
+                    match message_receiver.recv_timeout(remaining_time) {
+                        Ok(BatchMessage::ExportLog(log)) => {
+                            logs.push(log);
+                            if logs.len() == config.max_export_batch_size
+                                || last_export_time.elapsed() >= config.scheduled_delay
+                            {
+                                let _ = export_with_timeout_sync(
+                                    remaining_time,
+                                    exporter.as_mut(),
+                                    logs.split_off(0),
+                                    &mut last_export_time,
+                                );
+                            }
+                        }
+                        Ok(BatchMessage::ForceFlush(sender)) => {
+                            let result = export_with_timeout_sync(
+                                remaining_time,
+                                exporter.as_mut(),
+                                logs.split_off(0),
+                                &mut last_export_time,
+                            );
+                            let _ = sender.send(result);
+                        }
+                        Ok(BatchMessage::Shutdown(sender)) => {
+                            let result = export_with_timeout_sync(
+                                remaining_time,
+                                exporter.as_mut(),
+                                logs.split_off(0),
+                                &mut last_export_time,
+                            );
+                            let _ = sender.send(result);
+
+                            //
+                            // break out the loop and return from the current background thread.
+                            //
+                            break;
+                        }
+                        Ok(BatchMessage::SetResource(resource)) => {
+                            exporter.set_resource(&resource);
+                        }
+                        Err(RecvTimeoutError::Timeout) => {
+                            let _ = export_with_timeout_sync(
+                                remaining_time,
+                                exporter.as_mut(),
+                                logs.split_off(0),
+                                &mut last_export_time,
+                            );
+                        }
+                        Err(err) => {
+                            // TODO: this should not happen! Log the error and continue for now.
+                            otel_error!(
+                                name: "BatchLogProcessor.InternalError",
+                                error = format!("{}", err)
+                            );
+                        }
+                    }
+                }
+            })
+            .expect("Thread spawn failed."); //TODO: Handle thread spawn failure
+
+        // Return batch processor with link to worker
+        BatchLogProcessor {
+            message_sender,
+            handle: Mutex::new(Some(handle)),
+            forceflush_timeout: Duration::from_secs(5), // TODO: make this configurable
+            shutdown_timeout: Duration::from_secs(5),   // TODO: make this configurable
+            is_shutdown: AtomicBool::new(false),
+            dropped_logs_count: AtomicUsize::new(0),
+            max_queue_size,
+        }
+    }
+
+    /// Create a new batch processor builder
+    pub fn builder<E>(exporter: E) -> BatchLogProcessorBuilder<E>
+    where
+        E: LogExporter,
+    {
+        BatchLogProcessorBuilder {
+            exporter,
+            config: Default::default(),
+        }
+    }
+}
+
+#[allow(clippy::vec_box)]
+fn export_with_timeout_sync<E>(
+    _: Duration, // TODO, enforcing timeout in exporter.
+    exporter: &mut E,
+    batch: Vec<Box<(LogRecord, InstrumentationScope)>>,
+    last_export_time: &mut Instant,
+) -> ExportResult
+where
+    E: LogExporter + ?Sized,
+{
+    *last_export_time = Instant::now();
+
+    if batch.is_empty() {
+        return LogResult::Ok(());
+    }
+
+    let log_vec: Vec<(&LogRecord, &InstrumentationScope)> = batch
+        .iter()
+        .map(|log_data| (&log_data.0, &log_data.1))
+        .collect();
+    let export = exporter.export(LogBatch::new(log_vec.as_slice()));
+    let export_result = futures_executor::block_on(export);
+
+    match export_result {
+        Ok(_) => LogResult::Ok(()),
+        Err(err) => {
+            otel_error!(
+                name: "BatchLogProcessor.ExportError",
+                error = format!("{}", err)
+            );
+            LogResult::Err(err)
+        }
+    }
+}
+
+///
+/// A builder for creating [`BatchLogProcessor`] instances.
+///
+#[derive(Debug)]
+pub struct BatchLogProcessorBuilder<E> {
+    exporter: E,
+    config: BatchConfig,
+}
+
+impl<E> BatchLogProcessorBuilder<E>
+where
+    E: LogExporter + 'static,
+{
+    /// Set the BatchConfig for [`BatchLogProcessorBuilder`]
+    pub fn with_batch_config(self, config: BatchConfig) -> Self {
+        BatchLogProcessorBuilder { config, ..self }
+    }
+
+    /// Build a batch processor
+    pub fn build(self) -> BatchLogProcessor {
+        BatchLogProcessor::new(Box::new(self.exporter), self.config)
+    }
+}
+
+#[cfg(feature = "experimental_logs_batch_log_processor_with_async_runtime")]
+#[allow(clippy::large_enum_variant)]
+#[derive(Debug)]
+enum BatchMessageWithAsyncRuntime {
+    /// Export logs, usually called when the log is emitted.
+    ExportLog((LogRecord, InstrumentationScope)),
+    /// Flush the current buffer to the backend, it can be triggered by
+    /// pre configured interval or a call to `force_push` function.
+    Flush(Option<oneshot::Sender<ExportResult>>),
+    /// Shut down the worker thread, push all logs in buffer to the backend.
+    Shutdown(oneshot::Sender<ExportResult>),
+    /// Set the resource for the exporter.
+    SetResource(Arc<Resource>),
+}
+
+/// A [`LogProcessor`] that asynchronously buffers log records and reports
+/// them at a pre-configured interval.
+#[cfg(feature = "experimental_logs_batch_log_processor_with_async_runtime")]
+pub struct BatchLogProcessorWithAsyncRuntime<R: RuntimeChannel> {
+    message_sender: R::Sender<BatchMessageWithAsyncRuntime>,
+
+    // Track dropped logs - we'll log this at shutdown
+    dropped_logs_count: AtomicUsize,
+
+    // Track the maximum queue size that was configured for this processor
+    max_queue_size: usize,
+}
+
+#[cfg(feature = "experimental_logs_batch_log_processor_with_async_runtime")]
+impl<R: RuntimeChannel> Debug for BatchLogProcessorWithAsyncRuntime<R> {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        f.debug_struct("BatchLogProcessor")
+            .field("message_sender", &self.message_sender)
+            .finish()
+    }
+}
+
+#[cfg(feature = "experimental_logs_batch_log_processor_with_async_runtime")]
+impl<R: RuntimeChannel> LogProcessor for BatchLogProcessorWithAsyncRuntime<R> {
+    fn emit(&self, record: &mut LogRecord, instrumentation: &InstrumentationScope) {
+        let result = self
+            .message_sender
+            .try_send(BatchMessageWithAsyncRuntime::ExportLog((
+                record.clone(),
+                instrumentation.clone(),
+            )));
 
         // TODO - Implement throttling to prevent error flooding when the queue is full or closed.
         if let Err(err) = result {
@@ -184,7 +540,7 @@ impl<R: RuntimeChannel> LogProcessor for BatchLogProcessor<R> {
     fn force_flush(&self) -> LogResult<()> {
         let (res_sender, res_receiver) = oneshot::channel();
         self.message_sender
-            .try_send(BatchMessage::Flush(Some(res_sender)))
+            .try_send(BatchMessageWithAsyncRuntime::Flush(Some(res_sender)))
             .map_err(|err| LogError::Other(err.into()))?;
 
         futures_executor::block_on(res_receiver)
@@ -195,7 +551,7 @@ impl<R: RuntimeChannel> LogProcessor for BatchLogProcessor<R> {
     fn shutdown(&self) -> LogResult<()> {
         let (res_sender, res_receiver) = oneshot::channel();
         self.message_sender
-            .try_send(BatchMessage::Shutdown(res_sender))
+            .try_send(BatchMessageWithAsyncRuntime::Shutdown(res_sender))
             .map_err(|err| LogError::Other(err.into()))?;
 
         futures_executor::block_on(res_receiver)
@@ -207,12 +563,13 @@ impl<R: RuntimeChannel> LogProcessor for BatchLogProcessor<R> {
         let resource = Arc::new(resource.clone());
         let _ = self
             .message_sender
-            .try_send(BatchMessage::SetResource(resource));
+            .try_send(BatchMessageWithAsyncRuntime::SetResource(resource));
     }
 }
 
-impl<R: RuntimeChannel> BatchLogProcessor<R> {
-    pub(crate) fn new<E>(mut exporter: E, config: BatchConfig, runtime: R) -> Self
+#[cfg(feature = "experimental_logs_batch_log_processor_with_async_runtime")]
+impl<R: RuntimeChannel> BatchLogProcessorWithAsyncRuntime<R> {
+    pub(crate) fn new<E>(mut exporter: E, config: BatchConfig, runtime: R) -> Self 
     where
         E: LogExporter + Send + Sync + 'static,
     {
@@ -227,7 +584,7 @@ impl<R: RuntimeChannel> BatchLogProcessor<R> {
             let ticker = inner_runtime
                 .interval(config.scheduled_delay)
                 .skip(1) // The ticker is fired immediately, so we should skip the first one to align with the interval.
-                .map(|_| BatchMessage::Flush(None));
+                .map(|_| BatchMessageWithAsyncRuntime::Flush(None));
             let timeout_runtime = inner_runtime.clone();
             let mut logs = Vec::new();
             let mut messages = Box::pin(stream::select(message_receiver, ticker));
@@ -235,7 +592,7 @@ impl<R: RuntimeChannel> BatchLogProcessor<R> {
             while let Some(message) = messages.next().await {
                 match message {
                     // Log has finished, add to buffer of pending logs.
-                    BatchMessage::ExportLog(log) => {
+                    BatchMessageWithAsyncRuntime::ExportLog(log) => {
                         logs.push(log);
                         if logs.len() == config.max_export_batch_size {
                             let result = export_with_timeout(
@@ -256,7 +613,7 @@ impl<R: RuntimeChannel> BatchLogProcessor<R> {
                         }
                     }
                     // Log batch interval time reached or a force flush has been invoked, export current spans.
-                    BatchMessage::Flush(res_channel) => {
+                    BatchMessageWithAsyncRuntime::Flush(res_channel) => {
                         let result = export_with_timeout(
                             config.max_export_timeout,
                             //Arc::clone(&exporter ),
@@ -276,7 +633,7 @@ impl<R: RuntimeChannel> BatchLogProcessor<R> {
                         }
                     }
                     // Stream has terminated or processor is shutdown, return to finish execution.
-                    BatchMessage::Shutdown(ch) => {
+                    BatchMessageWithAsyncRuntime::Shutdown(ch) => {
                         let result = export_with_timeout(
                             config.max_export_timeout,
                             //Arc::clone(&exporter ),
@@ -297,22 +654,26 @@ impl<R: RuntimeChannel> BatchLogProcessor<R> {
                         break;
                     }
                     // propagate the resource
-                    BatchMessage::SetResource(resource) => {
+                    BatchMessageWithAsyncRuntime::SetResource(resource) => {
                         exporter.set_resource(&resource);
                     }
                 }
             }
         }));
         // Return batch processor with link to worker
-        BatchLogProcessor { message_sender }
+        BatchLogProcessorWithAsyncRuntime {
+            message_sender,
+            dropped_logs_count: AtomicUsize::new(0),
+            max_queue_size,
+        }
     }
 
     /// Create a new batch processor builder
-    pub fn builder<E>(exporter: E, runtime: R) -> BatchLogProcessorBuilder<E, R>
+    pub fn builder<E>(exporter: E, runtime: R) -> BatchLogProcessorWithAsyncRuntimeBuilder<E, R>
     where
         E: LogExporter,
     {
-        BatchLogProcessorBuilder {
+        BatchLogProcessorWithAsyncRuntimeBuilder {
             exporter,
             config: Default::default(),
             runtime,
@@ -320,6 +681,7 @@ impl<R: RuntimeChannel> BatchLogProcessor<R> {
     }
 }
 
+#[cfg(feature = "experimental_logs_batch_log_processor_with_async_runtime")]
 async fn export_with_timeout<E, R>(
     time_out: Duration,
     //exporter: Arc<Mutex<E>>,
@@ -354,6 +716,7 @@ where
 /// Batch log processor configuration.
 /// Use [`BatchConfigBuilder`] to configure your own instance of [`BatchConfig`].
 #[derive(Debug)]
+#[allow(dead_code)]
 pub struct BatchConfig {
     /// The maximum queue size to buffer logs for delayed processing. If the
     /// queue gets full it drops the logs. The default value of is 2048.
@@ -491,44 +854,31 @@ impl BatchConfigBuilder {
     }
 }
 
-/// A builder for creating [`BatchLogProcessor`] instances.
+/// A builder for creating [`BatchLogProcessorWithAsyncRuntime`] instances.
 ///
 #[derive(Debug)]
-pub struct BatchLogProcessorBuilder<E, R> {
+#[cfg(feature = "experimental_logs_batch_log_processor_with_async_runtime")]
+pub struct BatchLogProcessorWithAsyncRuntimeBuilder<E, R> {
     exporter: E,
     config: BatchConfig,
     runtime: R,
 }
 
-impl<E, R> BatchLogProcessorBuilder<E, R>
+#[cfg(feature = "experimental_logs_batch_log_processor_with_async_runtime")]
+impl<E, R> BatchLogProcessorWithAsyncRuntimeBuilder<E, R>
 where
     E: LogExporter + 'static,
     R: RuntimeChannel,
 {
-    /// Set the BatchConfig for [`BatchLogProcessorBuilder`]
+    /// Set the BatchConfig for [`BatchLogProcessorWithAsyncRuntimeBuilder`]
     pub fn with_batch_config(self, config: BatchConfig) -> Self {
-        BatchLogProcessorBuilder { config, ..self }
+        BatchLogProcessorWithAsyncRuntimeBuilder { config, ..self }
     }
 
     /// Build a batch processor
-    pub fn build(self) -> BatchLogProcessor<R> {
-        BatchLogProcessor::new(self.exporter, self.config, self.runtime)
+    pub fn build(self) -> BatchLogProcessorWithAsyncRuntime<R> {
+        BatchLogProcessorWithAsyncRuntime::new(self.exporter, self.config, self.runtime)
     }
-}
-
-/// Messages sent between application thread and batch log processor's work thread.
-#[allow(clippy::large_enum_variant)]
-#[derive(Debug)]
-enum BatchMessage {
-    /// Export logs, usually called when the log is emitted.
-    ExportLog((LogRecord, InstrumentationScope)),
-    /// Flush the current buffer to the backend, it can be triggered by
-    /// pre configured interval or a call to `force_push` function.
-    Flush(Option<oneshot::Sender<ExportResult>>),
-    /// Shut down the worker thread, push all logs in buffer to the backend.
-    Shutdown(oneshot::Sender<ExportResult>),
-    /// Set the resource for the exporter.
-    SetResource(Arc<Resource>),
 }
 
 #[cfg(all(test, feature = "testing", feature = "logs"))]
@@ -549,7 +899,6 @@ mod tests {
             },
             BatchConfig, BatchConfigBuilder, LogProcessor, LoggerProvider, SimpleLogProcessor,
         },
-        runtime,
         testing::logs::InMemoryLogExporter,
         Resource,
     };
@@ -561,6 +910,11 @@ mod tests {
     use std::sync::atomic::{AtomicUsize, Ordering};
     use std::sync::{Arc, Mutex};
     use std::time::Duration;
+
+    #[cfg(feature = "experimental_logs_batch_log_processor_with_async_runtime")]
+    use super::BatchLogProcessorWithAsyncRuntime;
+    #[cfg(feature = "experimental_logs_batch_log_processor_with_async_runtime")]
+    use crate::runtime;
 
     #[derive(Debug, Clone)]
     struct MockLogExporter {
@@ -698,8 +1052,7 @@ mod tests {
             (OTEL_BLRP_EXPORT_TIMEOUT, Some("2046")),
         ];
         temp_env::with_vars(env_vars.clone(), || {
-            let builder =
-                BatchLogProcessor::builder(InMemoryLogExporter::default(), runtime::Tokio);
+            let builder = BatchLogProcessor::builder(InMemoryLogExporter::default());
 
             assert_eq!(builder.config.max_export_batch_size, 500);
             assert_eq!(
@@ -719,8 +1072,7 @@ mod tests {
         env_vars.push((OTEL_BLRP_MAX_QUEUE_SIZE, Some("120")));
 
         temp_env::with_vars(env_vars, || {
-            let builder =
-                BatchLogProcessor::builder(InMemoryLogExporter::default(), runtime::Tokio);
+            let builder = BatchLogProcessor::builder(InMemoryLogExporter::default());
             assert_eq!(builder.config.max_export_batch_size, 120);
             assert_eq!(builder.config.max_queue_size, 120);
         });
@@ -735,8 +1087,8 @@ mod tests {
             .with_max_queue_size(4)
             .build();
 
-        let builder = BatchLogProcessor::builder(InMemoryLogExporter::default(), runtime::Tokio)
-            .with_batch_config(expected);
+        let builder =
+            BatchLogProcessor::builder(InMemoryLogExporter::default()).with_batch_config(expected);
 
         let actual = &builder.config;
         assert_eq!(actual.max_export_batch_size, 1);
@@ -773,8 +1125,7 @@ mod tests {
         let exporter = MockLogExporter {
             resource: Arc::new(Mutex::new(None)),
         };
-        let processor =
-            BatchLogProcessor::new(exporter.clone(), BatchConfig::default(), runtime::Tokio);
+        let processor = BatchLogProcessor::new(exporter.clone(),, BatchConfig::default());
         let provider = LoggerProvider::builder()
             .with_log_processor(processor)
             .with_resource(
@@ -789,7 +1140,10 @@ mod tests {
                     .build(),
             )
             .build();
-        tokio::time::sleep(Duration::from_secs(2)).await; // set resource in batch span processor is not blocking. Should we make it blocking?
+
+        // wait for the batch processor to process the resource.
+        tokio::time::sleep(Duration::from_millis(100)).await;
+
         assert_eq!(exporter.get_resource().unwrap().into_iter().count(), 5);
         let _ = provider.shutdown();
     }
@@ -801,8 +1155,7 @@ mod tests {
         let exporter = InMemoryLogExporterBuilder::default()
             .keep_records_on_shutdown()
             .build();
-        let processor =
-            BatchLogProcessor::new(exporter.clone(), BatchConfig::default(), runtime::Tokio);
+        let processor = BatchLogProcessor::new(exporter.clone(), BatchConfig::default());
 
         let mut record = LogRecord::default();
         let instrumentation = InstrumentationScope::default();
@@ -840,11 +1193,24 @@ mod tests {
     }
 
     #[tokio::test(flavor = "current_thread")]
-    #[ignore = "See issue https://github.com/open-telemetry/opentelemetry-rust/issues/1968"]
-    async fn test_batch_log_processor_shutdown_with_async_runtime_current_flavor_multi_thread() {
+    async fn test_batch_log_processor_shutdown_under_async_runtime_current_flavor_multi_thread() {
         let exporter = InMemoryLogExporterBuilder::default().build();
-        let processor =
-            BatchLogProcessor::new(exporter.clone(), BatchConfig::default(), runtime::Tokio);
+        let processor = BatchLogProcessor::new(Box::new(exporter.clone()), BatchConfig::default());
+
+        processor.shutdown().unwrap();
+    }
+
+    #[cfg(feature = "experimental_logs_batch_log_processor_with_async_runtime")]
+    #[tokio::test(flavor = "current_thread")]
+    #[ignore = "See issue https://github.com/open-telemetry/opentelemetry-rust/issues/1968"]
+    async fn test_batch_log_processor_with_async_runtime_shutdown_under_async_runtime_current_flavor_multi_thread(
+    ) {
+        let exporter = InMemoryLogExporterBuilder::default().build();
+        let processor = BatchLogProcessorWithAsyncRuntime::new(
+            exporter.clone(),
+            BatchConfig::default(),
+            runtime::Tokio,
+        );
 
         //
         // deadloack happens in shutdown with tokio current_thread runtime
@@ -855,33 +1221,21 @@ mod tests {
     #[tokio::test(flavor = "current_thread")]
     async fn test_batch_log_processor_shutdown_with_async_runtime_current_flavor_current_thread() {
         let exporter = InMemoryLogExporterBuilder::default().build();
-        let processor = BatchLogProcessor::new(
-            exporter.clone(),
-            BatchConfig::default(),
-            runtime::TokioCurrentThread,
-        );
-
+        let processor = BatchLogProcessor::new(exporter.clone(), BatchConfig::default());
         processor.shutdown().unwrap();
     }
 
     #[tokio::test(flavor = "multi_thread")]
     async fn test_batch_log_processor_shutdown_with_async_runtime_multi_flavor_multi_thread() {
         let exporter = InMemoryLogExporterBuilder::default().build();
-        let processor =
-            BatchLogProcessor::new(exporter.clone(), BatchConfig::default(), runtime::Tokio);
-
+        let processor = BatchLogProcessor::new(exporter.clone(), BatchConfig::default());
         processor.shutdown().unwrap();
     }
 
     #[tokio::test(flavor = "multi_thread")]
     async fn test_batch_log_processor_shutdown_with_async_runtime_multi_flavor_current_thread() {
         let exporter = InMemoryLogExporterBuilder::default().build();
-        let processor = BatchLogProcessor::new(
-            exporter.clone(),
-            BatchConfig::default(),
-            runtime::TokioCurrentThread,
-        );
-
+        let processor = BatchLogProcessor::new(exporter.clone(), BatchConfig::default());
         processor.shutdown().unwrap();
     }