+++ opentelemetry-http/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## vNext
 
 - Bump msrv to 1.75.0.
-
+- Add "internal-logs" feature flag (enabled by default), and emit internal logs.
 
 ## 0.27.0